VideoNeuralStyleTransfer/requirements.txt
tensorflow==2.4.1
tensorflow-hub==0.11.0
matplotlib==3.0.3
ipython==7.16.1
Pillow==8.1.0
Flask==1.1.2
Flask-Cors==3.0.10
configparser==5.0.1 
numpy==1.19.5
glob2==0.7
glob3==0.0.1
Werkzeug==1.0.1
torch
torchvision
